requirements.txt
annotated-types==0.7.0
anyio==4.7.0
APScheduler==3.11.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.15
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
httptools==0.6.4
idna==3.10
limits==3.14.1
packaging==24.2
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.3
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
wrapt==1.17.0
